+++ src/Contract/MappingRegistryInterface.php
@@ -4,18 +4,16 @@ declare(strict_types=1);
 
 namespace Sirix\ObjectMapper\Contract;
 
-use Sirix\ObjectMapper\Definition\MappingDefinition;
-
 interface MappingRegistryInterface
 {
     /**
      * @param class-string $source
      * @param class-string $target
      */
-    public function get(string $source, string $target): MappingDefinition;
+    public function get(string $source, string $target): MappingDefinitionInterface;
 
     /**
-     * @return iterable<MappingDefinition>
+     * @return iterable<MappingDefinitionInterface>
      */
     public function all(): iterable;
 }